Detect degenerate (i.e. all-zero) public keys (fixes #6)
All-zero public keys (i.e. Montgomery-u coordinates) have a scalar
multiplication of all-zero (i.e. themselves) regardless of the input
scalar.

It's somewhat debatable as to whether an explicit check should be added
for this case (the Curve25519 web site, for example, suggests avoiding
it), but I'm going to side with caution here and go ahead and add the
check.

This implementation performs the check in the high-level Ruby API,
ensuring that we don't have to depend on every last provider to
implement it.